back_end/pyproject.toml
[tool.poetry]
name = "back-end"
version = "0.1.0"
description = ""
authors = ["Madi <madi_gaziz@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework-simplejwt = "^5.5.1"
psycopg2 = "^2.9.11"
pillow = "^12.1.1"
python-dotenv = "^1.2.2"
google-api-python-client = "^2.191.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.15.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
